+++ src/v3/mod.rs
@@ -1,1 +1,1 @@
-mod lex;
+mod lexer;
